Move block lists to the policy so it is not public
