+++ src/policy/marksweepspace/native_ms/mod.rs
@@ -1,5 +1,7 @@
 mod block;
+mod block_list;
 mod global;
 
 pub use block::*;
+pub use block_list::*;
 pub use global::*;
